Move public-register plugin models from danceschool.register to danceschool.core (Step 3)
Extracted from danceschool/register/models.py (history traceable via git log
--find-copies-harder on the original file).

Changes:
- core/models.py: adds PublicRegisterNavPluginModel,
  PublicRegisterEventPluginModel, PublicRegisterEventPluginChoice
- core/migrations/0058_public_register_plugins.py: creates the three models
  under the 'core' app label (tables: core_publicregister*)
- register/migrations/0011_public_register_event_plugin.py: deleted; the
  register app no longer owns these models
- register/models.py: removes the three class definitions; re-exports them
  from core.models so register/cms_plugins.py continues to work unchanged
  until Step 4
