+++ danceschool/core/models.py
@@ -5082,3 +5082,197 @@ class RegisterEventLimitedModel(CMSPlugin):
 
     class Meta:
         abstract = True
+
+
+class PublicRegisterNavPluginModel(CMSPlugin):
+    '''
+    Model for the public register navigation bar plugin.  The plugin renders
+    a sticky Bootstrap navbar whose links are populated by JavaScript after
+    page load by reading the data-section-title attributes of
+    .public-register-section elements (produced by PublicRegisterEventPlugin).
+
+    No sections are stored here; this is purely a display/UX plugin.
+    '''
+
+    title = models.CharField(
+        _('Navbar brand text'), max_length=200, blank=True, default='',
+        help_text=_(
+            'Optional text displayed at the left edge of the navbar. '
+            'Leave blank to show navigation links only.'
+        )
+    )
+
+    def __str__(self):
+        return self.title or str(_('Public register navigation bar'))
+
+    class Meta:
+        verbose_name = _('Public register navigation bar')
+        verbose_name_plural = _('Public register navigation bars')
+
+
+class PublicRegisterEventPluginModel(RegisterEventLimitedModel):
+    '''
+    CMS plugin model for the public-facing registration page.  Provides
+    filterable event listings without at-the-door-specific options (payment
+    methods, requireFullRegistration, autoCheckIn).
+    '''
+
+    title = models.CharField(
+        _('Section title'), max_length=250, default=_('Upcoming Events'), blank=True
+    )
+
+    cssClasses = models.CharField(
+        _('Custom CSS classes'), max_length=250, null=True, blank=True,
+        help_text=_('Classes are applied to the surrounding &lt;div&gt;')
+    )
+
+    template = models.CharField(
+        _('Plugin template'), max_length=250, null=True, blank=True
+    )
+
+    def copy_relations(self, oldinstance):
+        super().copy_relations(oldinstance)
+        self.publicregistereventpluginchoice_set.all().delete()
+        for choice in oldinstance.publicregistereventpluginchoice_set.all():
+            choice.pk = None
+            choice.eventPlugin = self
+            choice.save()
+
+    def get_short_description(self):
+        return self.title or self.id
+
+    def save(self, *args, **kwargs):
+        needs_default_choice = (
+            not self.publicregistereventpluginchoice_set.exists() if self.pk else True
+        )
+        super().save(*args, **kwargs)
+        if needs_default_choice:
+            PublicRegisterEventPluginChoice.objects.create(eventPlugin=self)
+
+    class Meta:
+        permissions = (
+            (
+                'choose_custom_public_plugin_template',
+                _('Can enter a custom plugin template for public register plugins.')
+            ),
+        )
+
+
+class PublicRegisterEventPluginChoice(models.Model):
+    '''
+    Configuration for how PublicRegisterEventPluginModel renders registration
+    inputs.  Each instance produces a labelled number input per available role
+    (or a single "General admission" input when no roles are defined).
+    '''
+
+    SOLDOUT_CHOICES = [
+        ('D', _('Display with sold-out label')),
+        ('H', _('Hide sold-out choices')),
+    ]
+
+    eventPlugin = models.ForeignKey(
+        PublicRegisterEventPluginModel,
+        verbose_name=_('Plugin'),
+        on_delete=models.CASCADE,
+    )
+
+    optionLabel = models.CharField(
+        _('Label prefix'), max_length=100, blank=True, default='',
+        help_text=_(
+            'Optional prefix shown before the role name, e.g. "Sign up as". '
+            'Leave blank to show only the role name.'
+        )
+    )
+
+    soldOutRule = models.CharField(
+        _('Rule for sold-out choices'), max_length=1, default='D',
+        choices=SOLDOUT_CHOICES,
+    )
+
+    data = models.JSONField(
+        _('Additional data attached to registrations'), default=dict, blank=True,
+        help_text=_(
+            'Custom JSON stored with each registration produced by this choice. '
+            'This value is kept server-side and is never transmitted through '
+            'the browser, so it cannot be modified by users.'
+        )
+    )
+
+    order = models.PositiveSmallIntegerField(default=0, blank=False, null=False)
+
+    def addChoices(self, event):
+        '''
+        Return a list of choice dicts for the given event — one entry per
+        available role, or a single "General admission" entry when the event
+        has no roles defined.
+        '''
+        from math import ceil
+
+        choices = []
+
+        event_roles = list(
+            event.eventrole_set.filter(capacity__gt=0).select_related('role')
+        )
+
+        if event_roles:
+            roles_data = [
+                {'name': er.role.name, 'id': er.role.id, 'capacity': er.capacity}
+                for er in event_roles
+            ]
+        elif isinstance(event, Series):
+            try:
+                dtype_roles = list(
+                    event.classDescription.danceTypeLevel.danceType.roles.all()
+                )
+            except Exception:
+                dtype_roles = []
+            if dtype_roles:
+                per_role_cap = ceil(event.capacity / len(dtype_roles))
+                roles_data = [
+                    {'name': r.name, 'id': r.id, 'capacity': per_role_cap}
+                    for r in dtype_roles
+                ]
+            else:
+                roles_data = []
+        else:
+            roles_data = []
+
+        if roles_data:
+            count_rows = event.eventregistration_set.filter(
+                cancelled=False, dropIn=False, registration__final=True
+            ).values('role_id').annotate(count=Count('id'))
+            count_map = {row['role_id']: row['count'] for row in count_rows}
+        else:
+            roles_data = [{'name': None, 'id': None, 'capacity': event.capacity}]
+            count_map = {None: event.eventregistration_set.filter(
+                cancelled=False, dropIn=False, registration__final=True
+            ).count()}
+
+        for i, role in enumerate(roles_data):
+            num_registered = count_map.get(role['id'], 0)
+            capacity = role['capacity'] or 0
+            sold_out = num_registered >= capacity
+
+            if sold_out and self.soldOutRule == 'H':
+                continue
+
+            label = ' '.join(filter(None, [
+                self.optionLabel,
+                role['name'] or str(_('General admission')),
+            ]))
+
+            choices.append({
+                'label': label,
+                'price': event.pricingTier.onlinePrice if event.pricingTier else 0,
+                'roleName': role['name'],
+                'roleId': role['id'],
+                'numRegistered': num_registered,
+                'capacity': capacity,
+                'soldOut': sold_out,
+                'choiceId': 'pubchoice_{}_{}_{}'.format(event.id, self.id, i),
+            })
+
+        return choices
+
+    class Meta:
+        ordering = ['order']
